LayoutTests/streams/readablestream-start.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>

test(function()
{
    var isStartCalled = false;
    var source =
    { start:
        function(enqueue, close, error) {
            assert_equals(this, source);
            assert_equals(typeof enqueue, 'function');
            assert_equals(typeof close, 'function');
            assert_equals(typeof error, 'function');
            assert_array_equals(Object.getOwnPropertyNames(close), ['name', 'length']);
            isStartCalled = true;
        }
    };
    var rs = new ReadableStream(source);
    assert_true(isStartCalled);
}, 'ReadableStream start should be called with the proper parameters');

test(function()
{
    mashedError = new Error('potato');
    assert_throws(mashedError, function() {
        var rs = new ReadableStream(
        { start:
            function(enqueue, close, error) {
                throw mashedError;
            }
        });
    });
}, 'ReadableStream start should be able to throw');

test(function() {
    assert_throws(new TypeError(), function() {
        new ReadableStream({ start: 'potato'});
    });
}, 'ReadableStream constructor should get a function as start argument');

test(function()
{
    var rs = new ReadableStream(
    { start:
        function(enqueue, close, error) {
            close();
        }
    });
    assert_equals(rs.state, 'closed');
}, 'ReadableStream start should be able to close the stream');

test(function()
{
    var closeFunction;
    var rs = new ReadableStream(
    { start:
        function(enqueue, close, error) {
            closeFunction = close;
        }
    });
    assert_equals(rs.state, 'waiting');
    closeFunction();
    assert_equals(rs.state, 'closed');
}, 'Calling ReadableStream close function should transition synchronously to closed state');

test(function()
{
    var closeFunction;
    var rs = new ReadableStream(
    { start:
        function(enqueue, close, error) {
            closeFunction = close;
        }
    });
    assert_equals(rs.state, 'waiting');
    closeFunction();
    assert_equals(rs.state, 'closed');
    closeFunction();
    assert_equals(rs.state, 'closed');
}, 'Calling ReadableStream close function several times should have no effect');


</script>
